refactor: 패키지 변경 및 컨벤션에 맞춘 코드 리팩토링
Input, Output domain.data로 패키지 변경
RacingGame 랜덤값에 시드를 부여하도록 생성자 호출부 변경
