+++ src/test/java/racinggame/controller/InputTest.java
@@ -4,7 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import racinggame.controller.Input;
+import racinggame.domain.data.Input;
 
 public class InputTest {
