+++ src/test/java/racinggame/controller/OutputTest.java
@@ -4,7 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racinggame.controller.Output;
+import racinggame.domain.data.Output;
 
 public class OutputTest {
     Output output;
