+++ src/test/java/racinggame/domain/RacingGameTest.java
@@ -6,10 +6,9 @@ import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import racinggame.controller.Output;
+import racinggame.domain.data.Output;
 import racinggame.domain.car.Car;
-import racinggame.controller.Input;
-import racinggame.domain.RacingGame;
+import racinggame.domain.data.Input;
 
 import java.util.ArrayList;
 import java.util.List;
